server: move AuthenticationTLS test to e2e
