+++ test/e2e/tls/testdata/client-key.pem
